gatsby-node 에서 유틸 코드 분리 (허용 안됨)
